requirements.txt
Django==5.0.4
djangorestframework==3.15.1
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.0.4
django-celery-beat==2.6.0
cachetools==5.3.3
django-redis==5.4.0
flower==2.0.1
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.2.1
django-admin-rangefilter==0.12.4
pyTelegramBotAPI==4.18.1
django-debug-toolbar==4.4.2
python-dotenv